ICM-139 Explicit Docker image version and name in the init script
Currently, the Docker initscript has the image version and name in multiple parts of the code.
In order to make it easier to create the ZDD rollback job, we want to centralize it in a variable
that could, in the near future, be read from a file (one file per component).