China buildslaves should support both Docker repositories
CI buildslaves in China should be able to push images to the chinese
Docker repository (registry.bwtsi.cn) and to pull images from both
global (docker.tradeshift.net) and chinese Docker repositories.

In order to achieve that, we need to be able to use the appropriate
certificates for each Docker repository. As a consequence, `docker::cert`
was changed to allow more flexibility using the content or the target
(when symlinked) of the files.